User: Reorganize service modules
